fix: Catch dropbox upload errors to console

+++ src/eventhandlers/on-new-user-image.ts
@@ -25,16 +25,29 @@ export default async (event: any) => {
   try {
     await dropbox.uploadImage(event.data.val())
   } catch (error) {
-    if (error.error.error[".tag"] === "invalid_access_token") {
-      console.warn("User", event.params.userId, "seems to have disconnected Peppermint from their Dropbox. Removing their account");
-      await FirebaseClient.GET_INSTANCE().deleteUser(event.params.userId);
-      return;
+    // WTF dropbox... Could you nest this any deeper??
+    if (
+      error.error &&
+      error.error.error &&
+      error.error.error[".tag"] === "invalid_access_token"
+    ) {
+      console.warn(
+        "User",
+        event.params.userId,
+        "seems to have disconnected Peppermint from their Dropbox. Removing their account"
+      )
+      await FirebaseClient.GET_INSTANCE().deleteUser(event.params.userId)
+      return
     } else {
-      console.error("Failed to upload post", event.params.postId, "to Dropbox of user", event.params.userId)
+      console.error(
+        "Failed to upload post",
+        event.params.postId,
+        "to Dropbox of user",
+        event.params.userId
+      )
       throw error
     }
   }
-
 
   const Maintenance = iocContainer.get<Maintenance>(TYPES.Maintenance)
   await Maintenance.runForUser(event.params.userId)
